backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
better-profanity>=0.7.0
werkzeug>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# BART/T5 via HuggingFace Inference API
huggingface-hub>=0.20.0

# Local embeddings (sentence-transformers + FAISS)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Web search APIs
google-search-results>=2.4.2

# Visualization
graphviz>=0.20.0
plotly>=5.18.0
kaleido>=0.2.1

# Report generation
reportlab>=4.0.0
python-pptx>=0.6.21
